fix: Fix seeded random number generation
Also properly fixed a bunch of visual issues due to non-continuous use of the random generators (because there was only 1 random generator for the whole game, now every instance of Minecraft's `Random` class has its own `RandomGenerator`).
Enchantments work properly (and in a predictable way, exactly like in vanilla) according my testing as well now.
The codebase is a lot cleaner now as a result.